+++ src/components/common/UserShareCode.tsx
@@ -1,0 +1,15 @@
+import Button from "src/shared/components/button";
+
+const UserShareCode = () => {
+  return (
+    <>
+      <p>공유 코드예요!</p>
+      <p>아래 코드를 공유해보세요.</p>
+      <div>s8dg9sd9f9</div>
+      <Button>복사하기</Button>
+      <button>상대방의 코드를 알고 있어요!</button>
+    </>
+  );
+};
+
+export default UserShareCode;
